move updateDocumentFolderPath from Page to PageScreen
